+++ packages/engine/src/systems/generators/IntegerIncrement.ts
@@ -1,0 +1,11 @@
+export class IntegerIncrement {
+  #value = 0;
+
+  incr(): number {
+    return this.#value++;
+  }
+
+  clear(): void {
+    this.#value = 0;
+  }
+}