+++ packages/engine/test/mocks.ts
@@ -7,12 +7,12 @@ import { ActorComponent, Behavior } from "../src/index.ts";
 export function createActor(): {
   behaviors: Record<string, Behavior[]>;
   components: ActorComponent[];
-  gameInstance: ReturnType<typeof createGameInstance>;
+  world: ReturnType<typeof createWorld>;
 } {
   return {
     behaviors: {},
     components: [],
-    gameInstance: createGameInstance()
+    world: createWorld()
   };
 }
 
@@ -23,14 +23,14 @@ export function createTreeActor() {
   };
 }
 
-export function createGameInstance(): {
-  scene: {
+export function createWorld(): {
+  sceneManager: {
     componentsToBeStarted: ActorComponent[];
     tree: ReturnType<typeof createTreeActor>;
   };
 } {
   return {
-    scene: {
+    sceneManager: {
       componentsToBeStarted: [],
       tree: createTreeActor()
     }